Fork Tutf_function X11_UTF_32_to_charset_function for basic X11 vs XFT.
